Step 21: Migrate version module to Rust
- Replaced version.cpp with minimal stub (94 → 11 lines)
- Replaced version.hpp with Rust-backed implementation using FFI
- Version data and logic now in seqwish-rs/src/version.rs
- Updated seqwish_rs.h with version FFI functions
- Removed DEMO_version.hpp (demonstration complete)

All 79 Rust tests passing. C++ build successful.
Binary works correctly, showing version information.
